revert files that need more thought
